uclamp: Fix sched_setattr call
The struct sched_param passed to sched_setattr() for uclamp needs to
have a valid policy and priority set.

Currently, sched_data->prio is set to a special value
THREAD_PRIORITY_UNCHANGED when the phase does not declare a specific
priority, and sched_data->policy is set to "same". This is problematic
as the kernel expects a valid policy and priority.  We currently pass
the special values to the kernel, which makes the syscall fail.

Fix that by maintaining a valid policy and priority in
thread_data_t->curr_sched_data->{policy,prio} and using it where
appropriate.