ftrace: Autodetect location of tracing folder
If you don't provide a tracing location in the global config key section,
rt-app usually looks in /sys/kernel/debug for the tracing folder.

As part of this, switch the internals to use the full tracing path
instead of relying upon adding /tracing/ to the path of files used.

This patch allows rt-app to examine the content of /etc/mtab, and apply
the following ordering:

* If tracefs is mounted, use that location
* If tracefs is not found anywhere, then look for debugfs
  - If we use debugfs, append "/tracing" to the path
* If we find neither, just use /sys/kernel/debug/tracing

Using the default means we couldn't find tracefs or debugfs mounted
and so is likely to fail later if you are using ftrace, but should have
no impact if you're not tracing.

All of the path selection is overridden if you provide the config key.

This should allow rt-app to work on any platform, at least from tracing
POV.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>