rt-app: Fix broken calibration procedure
Calibration procedure is broken and will fail when values below 50 are
computed due to some rounding error. Even if that rounding error is
fixed, I could not find any justification for the adaptive sampling
method used, and the weighted average used gives as much weight to the
last sample as to all the previous samples combined, which does not
sound particularly justified either.

To fix that, modify the procedure to use a simple average of a number of
runs.