rt-app: Fix memload()
The call to memset() is optimised out by the compiler as the written
buffer is not read later by rt-app.

To fix this, replace the call to memset() with a functionally identical
piece of code relying only on `volatile` to inform the compiler that the
effects of the write need to be seen (by some external observer).

Signed-off-by: Petre Tudor <petre-ionut.tudor@arm.com>